backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
torch==2.2.0
torchvision==0.17.0
Pillow==10.2.0
numpy==1.26.3
opencv-python==4.9.0.80
